requirements.txt
flask
cloudevents
PyYAML
pyvmomi
python-dotenv